requirement.txt
streamlit
tensorflow
Pillow
gtts
googletrans==4.0.0-rc1
